controle-acesso/requirements.txt
Flask
Flask-SQLAlchemy
marshmallow-sqlalchemy
flask_marshmallow
psycopg2-binary
PyJWT